Enable plugin to read auth config at runtime
Marathon reads plugin configuration from the file supplied by "--plugin_conf"
command line option. To update permissions, marathon restart was required. To
avoid that, and to let user permissions confiuration be independent of plugin
configuration, this change separates, user permissions from a file specified
in a "plugin-conf.json" for key "permissions-conf-file". It is also possible
to configure how often to check for updates to the file, by default, it does
check every 30 seconds and if the file is modified, the new permissions are
read. This interval can be configured by setting the key in plugin-conf.json
"permissions-conf-file-check-interval-seconds" to a number, for example, 60.